METRICS-2701: revert SHA-1: 325c72eb538db347b923185ce9e31c60a748305c
the direction is NOT part of what makes a resource unique, and the previous change would break our APi with AppInsight which promises to use the same hash for the same resource.
Until Insight itself displays external resources this bug has no effect and therefore need not be fixed.

The only current effect is that when requesting metrics on an external resource we must specify the direction (even though the returned metrics will be an aggregation of both directions)