Update Spring Plugin Version Dependencies
This change updates the spring-* plugins to use the latest versions of their
respective Spring versions. The dependencies are interconnected and as such
this update covers all spring-* plugins.

spring-neo4j is temporarily disabled by this update due to incompatible changes
in spring-data-neo4j and requires a more indepth fix.